+++ src/librawspeed/decoders/DngDecoder.cpp
@@ -804,24 +804,25 @@ bool DngDecoder::decodeBlackLevels(const TiffIFD* raw) const {
     // We so not have enough to fill all individually, read a single and copy it
     float value = black_entry->getFloat();
 
-    if (value < std::numeric_limits<BlackType>::min() ||
-        value > std::numeric_limits<BlackType>::max())
+    if (static_cast<double>(value) < std::numeric_limits<BlackType>::min() ||
+        static_cast<double>(value) > std::numeric_limits<BlackType>::max())
       ThrowRDE("Error decoding black level");
 
     for (int y = 0; y < 2; y++) {
       for (int x = 0; x < 2; x++)
-        mRaw->blackLevelSeparate[y * 2 + x] = value;
+        mRaw->blackLevelSeparate[y * 2 + x] = implicit_cast<int>(value);
     }
   } else {
     for (int y = 0; y < 2; y++) {
       for (int x = 0; x < 2; x++) {
         float value = black_entry->getFloat(y * blackdim.x + x);
 
-        if (value < std::numeric_limits<BlackType>::min() ||
-            value > std::numeric_limits<BlackType>::max())
+        if (static_cast<double>(value) <
+                std::numeric_limits<BlackType>::min() ||
+            static_cast<double>(value) > std::numeric_limits<BlackType>::max())
           ThrowRDE("Error decoding black level");
 
-        mRaw->blackLevelSeparate[y * 2 + x] = value;
+        mRaw->blackLevelSeparate[y * 2 + x] = implicit_cast<int>(value);
       }
     }
   }
@@ -839,11 +840,12 @@ bool DngDecoder::decodeBlackLevels(const TiffIFD* raw) const {
     for (int i = 0; i < 4; i++) {
       const float value =
           black_sum[i >> 1] / static_cast<float>(mRaw->dim.y) * 2.0F;
-      if (value < std::numeric_limits<BlackType>::min() ||
-          value > std::numeric_limits<BlackType>::max())
+      if (static_cast<double>(value) < std::numeric_limits<BlackType>::min() ||
+          static_cast<double>(value) > std::numeric_limits<BlackType>::max())
         ThrowRDE("Error decoding black level");
 
-      if (__builtin_sadd_overflow(mRaw->blackLevelSeparate[i], value,
+      if (__builtin_sadd_overflow(mRaw->blackLevelSeparate[i],
+                                  implicit_cast<int>(value),
                                   &mRaw->blackLevelSeparate[i]))
         ThrowRDE("Integer overflow when calculating black level");
     }
@@ -861,11 +863,12 @@ bool DngDecoder::decodeBlackLevels(const TiffIFD* raw) const {
     for (int i = 0; i < 4; i++) {
       const float value =
           black_sum[i & 1] / static_cast<float>(mRaw->dim.x) * 2.0F;
-      if (value < std::numeric_limits<BlackType>::min() ||
-          value > std::numeric_limits<BlackType>::max())
+      if (static_cast<double>(value) < std::numeric_limits<BlackType>::min() ||
+          static_cast<double>(value) > std::numeric_limits<BlackType>::max())
         ThrowRDE("Error decoding black level");
 
-      if (__builtin_sadd_overflow(mRaw->blackLevelSeparate[i], value,
+      if (__builtin_sadd_overflow(mRaw->blackLevelSeparate[i],
+                                  implicit_cast<int>(value),
                                   &mRaw->blackLevelSeparate[i]))
         ThrowRDE("Integer overflow when calculating black level");
     }
